@@ -5,8 +5,9 @@ import android.content.Context;
 import android.location.Address;
 import android.location.Location;
 import android.os.Handler;
-import android.support.annotation.Nullable;
-import android.support.annotation.RequiresPermission;
+
+import androidx.annotation.Nullable;
+import androidx.annotation.RequiresPermission;
 
 import com.google.android.gms.common.api.Api;
 import com.google.android.gms.common.api.GoogleApiClient;
