@@ -1,6 +1,6 @@
 package pl.charmas.android.reactivelocation2.observables.geofence;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.google.android.gms.common.api.GoogleApiClient;
 import com.google.android.gms.common.api.ResultCallback;
