@@ -1,8 +1,9 @@
 package pl.charmas.android.reactivelocation2.sample.utils;
 
-import android.support.annotation.CheckResult;
-import android.support.annotation.NonNull;
 import android.widget.TextView;
+
+import androidx.annotation.CheckResult;
+import androidx.annotation.NonNull;
 
 import io.reactivex.Observable;
 
